[ITA-2533] Truncate logs from ActiveMqSender
